refactor: 구 Product 관련 로직 전부 삭제
