+++ src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.tobe.domain;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -8,4 +9,6 @@ public interface ProductRepository {
     Product save(Product product);
 
     Optional<Product> findById(UUID productId);
+
+    List<Product> findAll();
 }
