+++ src/main/java/kitchenpos/products/tobe/ui/ProductRestController.java
@@ -1,6 +1,7 @@
 package kitchenpos.products.tobe.ui;
 
 import java.net.URI;
+import java.util.List;
 import java.util.UUID;
 import javax.validation.Valid;
 import kitchenpos.products.tobe.application.ProductService;
@@ -8,6 +9,7 @@ import kitchenpos.products.tobe.dto.CreateProductRequest;
 import kitchenpos.products.tobe.dto.ModifyProductPriceRequest;
 import kitchenpos.products.tobe.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -35,5 +37,10 @@ public class ProductRestController {
     public ResponseEntity<ProductResponse> changePrice(@PathVariable UUID productId,
                                                        @RequestBody @Valid ModifyProductPriceRequest request) {
         return ResponseEntity.ok(productService.changePrice(productId, request));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ProductResponse>> findAll() {
+        return ResponseEntity.ok(productService.findAll());
     }
 }
