--- src/main/java/kitchenpos/products/ui/ProductRestController.java
@@ -1,24 +1,0 @@
-package kitchenpos.products.ui;
-
-import java.util.List;
-import kitchenpos.products.application.ProductService;
-import kitchenpos.products.domain.Product;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/products")
-@RestController
-public class ProductRestController {
-    private final ProductService productService;
-
-    public ProductRestController(final ProductService productService) {
-        this.productService = productService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<Product>> findAll() {
-        return ResponseEntity.ok(productService.findAll());
-    }
-}
