--- src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -1,68 +1,0 @@
-package kitchenpos.products.application;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.UUID;
-import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.domain.Product;
-import kitchenpos.products.domain.ProductRepository;
-import kitchenpos.products.infra.PurgomalumClient;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import static kitchenpos.Fixtures.menu;
-import static kitchenpos.Fixtures.menuProduct;
-import static kitchenpos.Fixtures.product;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-class ProductServiceTest {
-    private ProductRepository productRepository;
-    private MenuRepository menuRepository;
-    private PurgomalumClient purgomalumClient;
-    private ProductService productService;
-
-    @BeforeEach
-    void setUp() {
-        productRepository = new InMemoryProductRepository();
-        menuRepository = new InMemoryMenuRepository();
-        purgomalumClient = new FakePurgomalumClient();
-        productService = new ProductService(productRepository, menuRepository, purgomalumClient);
-    }
-
-    @DisplayName("상품의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        productRepository.save(product("후라이드", 16_000L));
-        productRepository.save(product("양념치킨", 16_000L));
-        final List<Product> actual = productService.findAll();
-        assertThat(actual).hasSize(2);
-    }
-
-    private Product createProductRequest(final String name, final long price) {
-        return createProductRequest(name, BigDecimal.valueOf(price));
-    }
-
-    private Product createProductRequest(final String name, final BigDecimal price) {
-        final Product product = new Product();
-        product.setName(name);
-        product.setPrice(price);
-        return product;
-    }
-
-    private Product changePriceRequest(final long price) {
-        return changePriceRequest(BigDecimal.valueOf(price));
-    }
-
-    private Product changePriceRequest(final BigDecimal price) {
-        final Product product = new Product();
-        product.setPrice(price);
-        return product;
-    }
-}
