+++ src/test/java/kitchenpos/products/tobe/domain/MockProductRepository.java
@@ -1,6 +1,8 @@
 package kitchenpos.products.tobe.domain;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
@@ -18,5 +20,10 @@ public class MockProductRepository implements ProductRepository {
     @Override
     public Optional<Product> findById(final UUID id) {
         return Optional.ofNullable(cache.get(id));
+    }
+
+    @Override
+    public List<Product> findAll() {
+        return new ArrayList<>(cache.values());
     }
 }
